+++ src/main/java/se/cygni/mss/tsv/model/Rating.java
@@ -8,7 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigInteger;
 
 @Entity
 @AllArgsConstructor
@@ -18,7 +17,7 @@ public class Rating {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private BigInteger rating_id;
+    private Integer rating_id;
     private String tconst;
     private String averageRating;
     private String numVotes;
